Change inheritance order
