+++ library/Zend/Form/Element/MonthSelect.php
@@ -10,11 +10,141 @@
 
 namespace Zend\Form\Element;
 
+use Zend\Form\Element;
+use Zend\Form\ElementPrepareAwareInterface;
+use Zend\Form\Form;
+use Zend\InputFilter\InputProviderInterface;
 use Zend\Validator\ValidatorInterface;
 use Zend\Validator\Regex as RegexValidator;
 
-class MonthSelect extends DateSelect
+class MonthSelect extends Element implements InputProviderInterface, ElementPrepareAwareInterface
 {
+    /**
+     * Select form element that contains values for month
+     *
+     * @var Select
+     */
+    protected $monthElement;
+
+    /**
+     * Select form element that contains values for year
+     *
+     * @var Select
+     */
+    protected $yearElement;
+
+    /**
+     * Min year to use for the select (default: current year - 100)
+     *
+     * @var int
+     */
+    protected $minYear;
+
+    /**
+     * Max year to use for the select (default: current year)
+     *
+     * @var int
+     */
+    protected $maxYear;
+
+    /**
+     * @var ValidatorInterface
+     */
+    protected $validator;
+
+
+    /**
+     * Constructor. Add two selects elements
+     *
+     * @param  null|int|string  $name    Optional name for the element
+     * @param  array            $options Optional options for the element
+     */
+    public function __construct($name = null, $options = array())
+    {
+        parent::__construct($name, $options);
+
+        $this->monthElement = new Select('month');
+        $this->yearElement = new Select('year');
+
+        $this->maxYear = date('Y');
+        $this->minYear = $this->maxYear - 100;
+    }
+
+    /**
+     * Accepted options for DateSelect:
+     * - min_year: min year to use in the year select
+     * - max_year: max year to use in the year select
+     *
+     * @param array|\Traversable $options
+     * @return DateSelect
+     */
+    public function setOptions($options)
+    {
+        parent::setOptions($options);
+
+        if (isset($options['min_year'])) {
+            $this->setMinYear($options['min_year']);
+        }
+
+        if (isset($options['max_year'])) {
+            $this->setMaxYear($options['max_year']);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Select
+     */
+    public function getMonthElement()
+    {
+        return $this->monthElement;
+    }
+
+    /**
+     * @return Select
+     */
+    public function getYearElement()
+    {
+        return $this->yearElement;
+    }
+
+    /**
+     * @param  int $minYear
+     * @return DateSelect
+     */
+    public function setMinYear($minYear)
+    {
+        $this->minYear = $minYear;
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getMinYear()
+    {
+        return $this->minYear;
+    }
+
+    /**
+     * @param  int $maxYear
+     * @return DateSelect
+     */
+    public function setMaxYear($maxYear)
+    {
+        $this->maxYear = $maxYear;
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getMaxYear()
+    {
+        return $this->maxYear;
+    }
+
     /**
      * @param mixed $value
      * @return void|\Zend\Form\Element
@@ -23,6 +153,19 @@ class MonthSelect extends DateSelect
     {
         $this->monthElement->setValue($value['month']);
         $this->yearElement->setValue($value['year']);
+    }
+
+    /**
+     * Prepare the form element (mostly used for rendering purposes)
+     *
+     * @param Form $form
+     * @return mixed
+     */
+    public function prepareElement(Form $form)
+    {
+        $name = $this->getName();
+        $this->monthElement->setName($name . '[month]');
+        $this->yearElement->setName($name . '[year]');
     }
 
     /**
