+++ library/Zend/Form/View/Helper/FormMonthSelect.php
@@ -10,8 +10,10 @@
 
 namespace Zend\Form\View\Helper;
 
+use DateTime;
+use IntlDateFormatter;
+use Locale;
 use Zend\Form\ElementInterface;
-use Zend\Form\View\Helper\FormDateSelect as FormDateSelectHelper;
 use Zend\Form\Exception;
 
 /**
@@ -19,12 +21,41 @@ use Zend\Form\Exception;
  * @package    Zend_Form
  * @subpackage View
  */
-class FormMonthSelect extends FormDateSelectHelper
+class FormMonthSelect extends AbstractHelper
 {
+    /**
+     * FormSelect helper
+     *
+     * @var FormSelect
+     */
+    protected $selectHelper;
+
+    /**
+     * Date formatter to use
+     *
+     * @var int
+     */
+    protected $dateType = IntlDateFormatter::LONG;
+
+    /**
+     * Pattern to use for Date rendering
+     *
+     * @var string
+     */
+    protected $pattern;
+
+    /**
+     * Locale to use
+     *
+     * @var string
+     */
+    protected $locale;
+
     /**
      * Render a month element that is composed of two selects
      *
      * @param \Zend\Form\ElementInterface $element
+     * @throws \Zend\Form\Exception\DomainException
      * @return string
      */
     public function render(ElementInterface $element)
@@ -62,6 +93,177 @@ class FormMonthSelect extends FormDateSelectHelper
         );
 
         return $markup;
+    }
+
+    /**
+     * Invoke helper as function
+     *
+     * Proxies to {@link render()}.
+     *
+     * @param \Zend\Form\ElementInterface $element
+     * @param int                         $dateType
+     * @param null|string                 $locale
+     * @return FormDateSelect
+     */
+    public function __invoke(ElementInterface $element = null, $dateType = IntlDateFormatter::LONG, $locale = null)
+    {
+        if (!$element) {
+            return $this;
+        }
+
+        $this->setDateType($dateType);
+
+        if ($locale !== null) {
+            $this->setLocale($locale);
+        }
+
+        return $this->render($element);
+    }
+
+    /**
+     * @return string
+     */
+    public function getPattern()
+    {
+        if ($this->pattern === null) {
+            $intl           = new IntlDateFormatter($this->getLocale(), $this->dateType, IntlDateFormatter::NONE);
+            $this->pattern  = $intl->getPattern();
+        }
+
+        return $this->pattern;
+    }
+
+    /**
+     * Parse the pattern
+     *
+     * @return array
+     */
+    protected function parsePattern()
+    {
+        $pattern    = $this->getPattern();
+        $pregResult = preg_split('/([ -,.\/]+)/', $pattern, -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY);
+
+        $result = array();
+        foreach ($pregResult as $value) {
+            if (stripos($value, 'd') !== false) {
+                $result['day'] = $value;
+            } elseif (stripos($value, 'm') !== false) {
+                $result['month'] = $value;
+            } elseif (stripos($value, 'y') !== false) {
+                $result['year'] = $value;
+            } else {
+                $result[] = $value;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * @param  int $dateType
+     * @return FormDateSelect
+     */
+    public function setDateType($dateType)
+    {
+        // The FULL format uses values that are not used
+        if ($dateType === IntlDateFormatter::FULL) {
+            $dateType = IntlDateFormatter::LONG;
+        }
+
+        $this->dateType = $dateType;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getDateType()
+    {
+        return $this->dateType;
+    }
+
+    /**
+     * @param  string $locale
+     * @return FormDateSelect
+     */
+    public function setLocale($locale)
+    {
+        $this->locale = $locale;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getLocale()
+    {
+        if ($this->locale === null) {
+            $this->locale = Locale::getDefault();
+        }
+
+        return $this->locale;
+    }
+
+    /**
+     * Create a key => value options for months
+     *
+     * @param string $pattern Pattern to use for months
+     * @return array
+     */
+    public function getMonthsOptions($pattern)
+    {
+        $keyFormatter   = new IntlDateFormatter($this->getLocale(), null, null, null, null, 'MM');
+        $valueFormatter = new IntlDateFormatter($this->getLocale(), null, null, null, null, $pattern);
+        $date           = new DateTime('1970-01-01');
+
+        $result = array();
+        for ($month = 1; $month <= 12; $month++) {
+            $key   = $keyFormatter->format($date);
+            $value = $valueFormatter->format($date);
+            $result[$key] = $value;
+
+            $date->modify('+1 month');
+        }
+
+        return $result;
+    }
+
+    /**
+     * Create a key => value options for years
+     * NOTE: we don't use a pattern for years, as years written as two digits can lead to hard to
+     * read date for users, so we only use four digits years
+     *
+     * @param int $minYear
+     * @param int $maxYear
+     * @return array
+     */
+    public function getYearsOptions($minYear, $maxYear)
+    {
+        $result = array();
+        for ($i = $maxYear; $i >= $minYear; --$i) {
+            $result[$i] = $i;
+        }
+
+        return $result;
+    }
+
+    /**
+     * Retrieve the FormSelect helper
+     *
+     * @return FormRow
+     */
+    protected function getSelectElementHelper()
+    {
+        if ($this->selectHelper) {
+            return $this->selectHelper;
+        }
+
+        if (method_exists($this->view, 'plugin')) {
+            $this->selectHelper = $this->view->plugin('formselect');
+        }
+
+        return $this->selectHelper;
     }
 }
